Factor out `BlockFetchServerEventTrace` to its own module Rename `ToObjectOrphans` to `ConsensusToObjectOrphans`
